Move metrics code
